Rename alexnet_classifying.py to alexnet_classify.py
Change the name from alexnet_classifying.py to alexnet_classify.py.  Users need to run the script to download the designated data in the first time. In the second or more times, it is more adaptable for users to run  alexnet_classify.py_simple without needing repeated download of data.