Add the script to facilitate the visible execution and tensorboard rendering.
Add the classification script in the .ipynb format in order to vividly address the running of the line of code. The script allows users to download and parse the image data. After downloading the image data, users can choose to use the lightweight alexnet_classify_simple.ipynb to run the data